+++ lbr_bringup/lbr_bringup/description.py
@@ -1,7 +1,6 @@
 from typing import Dict, List, Optional, Union
 
-from launch.actions import DeclareLaunchArgument, IncludeLaunchDescription
-from launch.launch_description_sources import PythonLaunchDescriptionSource
+from launch.actions import DeclareLaunchArgument
 from launch.substitutions import (
     Command,
     FindExecutable,
@@ -10,48 +9,6 @@ from launch.substitutions import (
 )
 from launch_ros.actions import Node
 from launch_ros.substitutions import FindPackageShare
-
-
-class GazeboMixin:
-    @staticmethod
-    def include_gazebo(**kwargs) -> IncludeLaunchDescription:
-        return IncludeLaunchDescription(
-            PythonLaunchDescriptionSource(
-                PathJoinSubstitution(
-                    [
-                        FindPackageShare("gazebo_ros"),
-                        "launch",
-                        "gazebo.launch.py",
-                    ]
-                )
-            ),
-            **kwargs,
-        )
-
-    @staticmethod
-    def node_spawn_entity(
-        robot_name: Optional[Union[LaunchConfiguration, str]] = LaunchConfiguration(
-            "robot_name", default="lbr"
-        ),
-        tf: List[float] = [0.0, 0.0, 0.0, 0.0, 0.0, 0.0],
-        **kwargs,
-    ) -> Node:
-        label = ["-x", "-y", "-z", "-R", "-P", "-Y"]
-        tf = [str(x) for x in tf]
-        return Node(
-            package="gazebo_ros",
-            executable="spawn_entity.py",
-            arguments=[
-                "-topic",
-                "robot_description",
-                "-entity",
-                robot_name,
-            ]
-            + [item for pair in zip(label, tf) for item in pair],
-            output="screen",
-            namespace=robot_name,
-            **kwargs,
-        )
 
 
 class LBRDescriptionMixin:
@@ -66,12 +23,10 @@ class LBRDescriptionMixin:
         port_id: Optional[Union[LaunchConfiguration, str]] = LaunchConfiguration(
             "port_id", default="30200"
         ),
-        sim: Optional[Union[LaunchConfiguration, bool]] = LaunchConfiguration(
-            "sim", default="true"
+        mode: Optional[Union[LaunchConfiguration, bool]] = LaunchConfiguration(
+            "mode", default="mock"
         ),
     ) -> Dict[str, str]:
-        if type(sim) is bool:
-            sim = "true" if sim else "false"
         robot_description = {
             "robot_description": Command(
                 [
@@ -90,8 +45,8 @@ class LBRDescriptionMixin:
                     robot_name,
                     " port_id:=",
                     port_id,
-                    " sim:=",
-                    sim,
+                    " mode:=",
+                    mode,
                 ]
             )
         }
@@ -124,11 +79,16 @@ class LBRDescriptionMixin:
         )
 
     @staticmethod
-    def arg_sim(default_value: str = "true") -> DeclareLaunchArgument:
+    def arg_mode(default_value: str = "mock") -> DeclareLaunchArgument:
         return DeclareLaunchArgument(
-            name="sim",
+            name="mode",
             default_value=default_value,
-            description="Whether to use the simulation or not.",
+            description="The mode to launch in.",
+            choices=[
+                "mock",
+                "hardware",
+                "gazebo",
+            ],
         )
 
     @staticmethod
@@ -140,8 +100,8 @@ class LBRDescriptionMixin:
         return {"port_id": LaunchConfiguration("port_id", default="30200")}
 
     @staticmethod
-    def param_sim() -> Dict[str, LaunchConfiguration]:
-        return {"sim": LaunchConfiguration("sim", default="true")}
+    def param_mode() -> Dict[str, LaunchConfiguration]:
+        return {"mode": LaunchConfiguration("mode", default="mock")}
 
     @staticmethod
     def node_static_tf(
@@ -150,6 +110,8 @@ class LBRDescriptionMixin:
         child: Optional[Union[LaunchConfiguration, str]] = None,
         **kwargs,
     ) -> Node:
+        if len(tf) != 6:
+            raise ValueError("tf must be a list of 6 floats.")
         label = ["--x", "--y", "--z", "--roll", "--pitch", "--yaw"]
         tf = [str(x) for x in tf]
         return Node(
@@ -162,54 +124,6 @@ class LBRDescriptionMixin:
                 parent,
                 "--child-frame-id",
                 child,
-            ],
-            **kwargs,
-        )
-
-
-class RVizMixin:
-    @staticmethod
-    def arg_rviz_config_pkg(
-        default_value: str = "lbr_description",
-    ) -> DeclareLaunchArgument:
-        return DeclareLaunchArgument(
-            name="rviz_config_pkg",
-            default_value=default_value,
-            description="The RViz configuration file.",
-        )
-
-    @staticmethod
-    def arg_rviz_config(
-        default_value: str = "config/config.rviz",
-    ) -> DeclareLaunchArgument:
-        return DeclareLaunchArgument(
-            name="rviz_config",
-            default_value=default_value,
-            description="The RViz configuration file.",
-        )
-
-    @staticmethod
-    def node_rviz(
-        rviz_config_pkg: Optional[
-            Union[LaunchConfiguration, str]
-        ] = LaunchConfiguration("rviz_config_pkg", default="lbr_description"),
-        rviz_config: Optional[Union[LaunchConfiguration, str]] = LaunchConfiguration(
-            "rviz_config", default="config/config.rviz"
-        ),
-        **kwargs,
-    ) -> Node:
-        return Node(
-            package="rviz2",
-            executable="rviz2",
-            name="rviz2",
-            arguments=[
-                "-d",
-                PathJoinSubstitution(
-                    [
-                        FindPackageShare(rviz_config_pkg),
-                        rviz_config,
-                    ]
-                ),
             ],
             **kwargs,
         )