+++ lbr_fri_ros2/include/lbr_fri_ros2/formatting.hpp
@@ -76,5 +76,5 @@ struct EnumMaps {
     }
   };
 };
-} // end of namespace lbr_fri_ros2
+} // namespace lbr_fri_ros2
 #endif // LBR_FRI_ROS2__ENUM_MAPS_HPP