+++ lbr_fri_ros2/src/interfaces/torque_command.cpp
@@ -48,4 +48,4 @@ void TorqueCommandInterface::buffered_command_to_fri(fri_command_t_ref command,
   command.setJointPosition(command_.joint_position.data());
   command.setTorque(command_.torque.data());
 }
-} // end of namespace lbr_fri_ros2
+} // namespace lbr_fri_ros2