+++ lbr_ros2_control/src/controllers/lbr_torque_command_controller.cpp
@@ -1,4 +1,4 @@
-#include "lbr_ros2_control/lbr_torque_command_controller.hpp"
+#include "lbr_ros2_control/controllers/lbr_torque_command_controller.hpp"
 
 namespace lbr_ros2_control {
 LBRTorqueCommandController::LBRTorqueCommandController()
@@ -102,7 +102,7 @@ void LBRTorqueCommandController::clear_command_interfaces_() {
   joint_position_command_interfaces_.clear();
   torque_command_interfaces_.clear();
 }
-} // end of namespace lbr_ros2_control
+} // namespace lbr_ros2_control
 
 #include "pluginlib/class_list_macros.hpp"